requirements.txt
# Core RL and Environment
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
carl-bench>=1.1.0

# Stable Baselines3 for PPO/SAC
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# Data handling and visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration management
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
einops>=0.7.0
